requirements.txt
# Core Web Framework
flask>=2.3.0,<3.0.0
flask-sqlalchemy>=3.0.0,<4.0.0

# AI/ML Dependencies
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0

# Optional ML Libraries
peft>=0.4.0

# Database
sqlite3  # Built-in with Python

# Security (recommended)
flask-wtf>=1.1.0
flask-bcrypt>=1.0.0

# Development Tools (optional)
python-dotenv>=1.0.0

